Revision to Improve Testability
Significant changes were made to the EndpointRequestor so that the shared URLSession is completely encapsulated in a single class so that it could swapped out with a mock version that loads and returns data from test vector files saved locally in the file system or app bundle. Test Vector capturing utilities were also including for capturing test vectors. An attribute is included for turning locally sourced data debugging on and off.